Dev (#44)
* The color fields in action 10014 are now vector fields.
This fixes recoloring of this action and greatly simplifies the class for it.

* Improved property minification.
Now, if all keyframes in a sequence property or component sequence property are equivalent, the property can be minified to a constant property.

* The recoloring update.
This includes a ton of things related to recoloring.
It also includes bug fixes and improvements to the anyValueSum and anyValueMult functions.

* Improvements and bug fixes for the color palette-related functions.
- Replaced the 'average' parameter for the palette generation functions with a 'mode' parameter that can be set to one of three values:
  - PaletteMode.First: Same as average=false before.
  - PaletteMode.Average: Same as average=true before.
  - PaletteMode.Random: Adds an entry for each set of colors for each slot in the palette. The Node.recolor method will pick random entries in the palette's slot for each effect.
- The Node.recolor method will now generate fallback entries for missing slots of the given palette based on the existing slots. This means it should be better at recoloring effects when the palette is missing many of the slots. The method no longer warns about missing entries due to this change.
- The common particle slot in color palettes have been split into 4 different ones based on the blend mode used. This should make some recolors more accurate to the sampled effect.
- Bloom strength is now handled slightly differently to prevent the effect from becoming too bright in some cases.

* Added an FXRUtility namespace for utility functions.
Currently only has two functions: line and text.
The line function is useful for drawing lines between 3D points.
The text function generates nodes that form characters without needing any textures. It creates a lot of nodes, so it's not the most optimized way of doing this, but it works really well for debugging.

* Fixed links to functions inside namespaces in the docs website.

* Documented the hex template tag.

* Fixed point lights being way too bright in DS3.

* Added documentation for the FXRUtility namespace and its functions.

* Fixed more broken function links on the docs website.

* Added clone and scaleRateOfTime methods.
FXR objects, nodes, effects, and actions now have a clone method.
Nodes and DataActions now have a scaleRateOfTime method.
When writing for DS3, the rate of time from action 10500 is now automatically applied to all nodes. States still need to have this applied to them, but that will be in a future commit as I'm out of time right now.

* Writing to DS3 now scales the rate of time for states as well.
This was missing from the previous commit. With this, the rate of time should work properly for DS3 just like it does for the other games.

* Updated changelog.

* Throw when writing invalid FXR ID + FXR.name getter.

* Updated readme and Node.js setup guide.

* Cleaned up the Recolor.standardBlend function a bit.

* Fixed the Recolor.hueShift function always returning cyan-ish colors.

* Allow Vector3s to be used for more of the recolor functions.

* Fixed prototype method links not being redirected properly in docs site.

* Changed the fields in the node transform actions to vector fields.

* Updated readme, changelog, and examples.

* Added methods for getting all unique color values in nodes/actions.

* Removed dependency on Node.js Buffers when writing to a file.
Still requires an import for 'node:fs/promises', but that can be aliased in Vite for example if needed in the browser.

* Avoid the flashbang when going to the /q path on the docs site.

* Redirect to the home page when there no matches on the docs site's /q.

* Added sections to the readme and the Node.js guide about the playground.

* Updated the creation example in the readme to use fxr.name.

* Added links to open the two readme examples in the playground.

* 13.0.0

* Tagged v13.0.0 in the changelog.